Create folders DAO and DTO
