+++ src/public/css/login.css
@@ -3,6 +3,23 @@ body{
     padding: 0;
     box-sizing: border-box;
 }
+
+.navBar{
+    display: flex;
+    justify-content: center;
+    gap: 10%;
+    align-items: center;
+    padding: 10px;
+    background-color: #000000;
+    height: 10vh;
+}
+.navBar a{
+    text-decoration: none;
+    color: rgb(255, 255, 255);
+    font-size: 20px;
+    cursor: pointer;
+}
+
 .loginContainer{
     display: flex;
     flex-direction: column;
