+++ src/public/js/profile.js
@@ -1,0 +1,4 @@
+document.querySelector("#logout").addEventListener("click", async (e) => {
+    localStorage.removeItem("cartID");
+    
+})
