+++ src/views/cart.handlebars
@@ -1,5 +1,5 @@
 <nav class="navBar">
-    <a class="cart">Cart</a>
+    <a class="cart" id="{{findCart}}">Cart</a>
     <a href="/products">Products</a>
     <a href="/createProducts">Create Products</a>
 </nav>
