+++ src/views/productDetails.handlebars
@@ -1,5 +1,5 @@
 <nav class="navBar">
-    <a href="/cart">Cart</a>
+    <a id="cart">Cart</a>
     <a href="/products">Products</a>
     <a href="/createProducts">Create Products</a>
 </nav>
